Airflow DataPipelines/requirements.txt
geopy==2.3.0
pandas==1.3.5
pyarrow==8.0.0
apache-airflow-providers-google==8.10.0 
google-cloud-storage==2.7.0
gcsfs==2022.11.0